// FrontEnd/FrontEnd/src/app/models/Educacion.ts
import { Usuarios } from "./Usuarios";

export class Educacion{
    idEdu?: number | undefined;
	iconoIns: string | undefined;
	dirIcono: string | undefined;
	institucion: string | undefined;
	inicio: number | undefined;
	fin: number | undefined;
	titulo: string | undefined;
    user: Usuarios | undefined;
}